Use `f64` instead of `usize` for fragment widths

+++ fuzz/fuzz_targets/wrap_first_fit.rs
@@ -4,21 +4,21 @@ use libfuzzer_sys::fuzz_target;
 use textwrap::core;
 use textwrap::wrap_algorithms::wrap_first_fit;
 
-#[derive(Arbitrary, Debug, Eq, PartialEq)]
+#[derive(Arbitrary, Debug, PartialEq)]
 struct Word {
-    width: usize,
-    whitespace_width: usize,
-    penalty_width: usize,
+    width: f64,
+    whitespace_width: f64,
+    penalty_width: f64,
 }
 
 #[rustfmt::skip]
 impl core::Fragment for Word {
-    fn width(&self) -> usize { self.width }
-    fn whitespace_width(&self) -> usize { self.whitespace_width }
-    fn penalty_width(&self) -> usize { self.penalty_width }
+    fn width(&self) -> f64 { self.width }
+    fn whitespace_width(&self) -> f64 { self.whitespace_width }
+    fn penalty_width(&self) -> f64 { self.penalty_width }
 }
 
-fuzz_target!(|input: (usize, Vec<Word>)| {
+fuzz_target!(|input: (f64, Vec<Word>)| {
     let width = input.0;
     let words = input.1;
     let _ = wrap_first_fit(&words, &[width]);
